Move tests, remove unused fixtures
